[fix]: view 분리 및 info 수정
