+++ Todoary-iOS/Todoary/Data/DataManager/Summary/GetCalendataManager.swift
@@ -25,7 +25,9 @@ class GetCalendataManager {
                     switch result.code {
                     case 1000:
                         print("캘린더조회성공")
-                        HomeViewController.successAPI_calendar(result.result)
+                        if let vc = viewController as? HomeViewController{
+                            vc.successAPI_calendar(result.result)
+                        }
                     case 2005:
                         print("유효하지 않은 회원정보입니다")
                     case 2010:
