+++ Todoary-iOS/Todoary/Data/DataManager/User/Profile/ProfileDataManager.swift
@@ -18,11 +18,11 @@ class ProfileDataManager {
                 switch result.code {
                 case 1000:
                     print("프로필수정성공")
-                    viewController.nickNameNotice.isHidden = true
+                    viewController.mainView.nickNameNotice.isHidden = true
                     viewController.navigationController?.popViewController(animated: true)
                 case 2032:
                     print("중복된 닉네임입니다")
-                    viewController.nickNameNotice.isHidden = false
+                    viewController.mainView.nickNameNotice.isHidden = false
                 default:
                     print(result.message)
                 }
