+++ Todoary-iOS/Todoary/Presentation/ViewControllers/User/Password/UserPassword/PwFindViewController.swift
@@ -39,158 +39,51 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
         }
     }
     
-    //MARK: - UIComponenets
-    
-    //id
-    let idTitle = UILabel().then{
-        $0.text = "아이디"
-        $0.textColor = .headline
-        $0.font = UIFont.nbFont(type: .subtitle)
-        $0.addLetterSpacing(spacing: 0.32)
-    }
-    
-    let idTf = UITextField().then{
-        $0.placeholder = "이메일을 입력해주세요"
-        $0.setPlaceholderColor()
-        $0.font = UIFont.nbFont(type: .body2)
-        $0.textFieldTypeSetting(type: .body1)
-        $0.returnKeyType = .next
-        $0.enablesReturnKeyAutomatically = true
-    }
-    
-    let idNoticeLb = UILabel().then{
-        $0.text = "*가입시 등록했던 이메일을 입력해주세요 "
-        $0.textColor = .todoaryGrey
-        $0.addLetterSpacing(spacing: 0.36)
-        $0.font = UIFont.nbFont(type: .sub1)
-    }
-    
-    let idBorderLine = UIView().then{
-        $0.backgroundColor = .todoaryGrey
-    }
-    
-    let idCertificationBtn = UIButton().then{
-        $0.setTitle("인증하기", for: .normal)
-        $0.setTitleColor(.white, for: .normal)
-        $0.titleLabel?.font = UIFont.nbFont(type: .subButton)
-        $0.buttonTypeSetting(type: .subButton)
-        $0.backgroundColor = .buttonColor
-        $0.layer.cornerRadius = 22/2
-        $0.addTarget(self, action: #selector(idCertificationBtnDidTap), for: .touchUpInside)
-    }
-    
-    //인증코드
-    let certificationTitle = UILabel().then{
-        $0.text = "인증코드 입력"
-        $0.textColor = .headline
-        $0.font = UIFont.nbFont(type: .subtitle)
-        $0.addLetterSpacing(spacing: 0.32)
-        
-    }
-
-    let certificationTf = UITextField().then{
-        $0.font = UIFont.nbFont(type: .body2)
-        $0.returnKeyType = .next
-        $0.enablesReturnKeyAutomatically = true
-    }
-
-    let certificationBorderLine = UIView().then{
-        $0.backgroundColor = .todoaryGrey
-    }
-    
-    let certificationOkBtn = UIButton().then{
-        $0.setTitle("확인", for: .normal)
-        $0.setTitleColor(.white, for: .normal)
-        $0.titleLabel?.font = UIFont.nbFont(type: .subButton)
-        $0.buttonTypeSetting(type: .subButton)
-        $0.backgroundColor = .buttonColor
-        $0.layer.cornerRadius = 22/2
-        $0.addTarget(self, action: #selector(certificationOkBtnDidTap), for: .touchUpInside)
-    }
-
-    //pw
-    let pwTitle = UILabel().then{
-        $0.text = "비밀번호 재설정"
-        $0.textColor = .headline
-        $0.font = UIFont.nbFont(type: .subtitle)
-        $0.addLetterSpacing(spacing: 0.32)
-    }
-
-    let pwTf = UITextField().then{
-        $0.placeholder = "영문, 숫자 포함 8자리 이상"
-        $0.isSecureTextEntry = true
-        $0.setPlaceholderColor()
-        $0.font = UIFont.nbFont(type: .body2)
-        $0.textFieldTypeSetting(type: .body1)
-        $0.returnKeyType = .next
-        $0.enablesReturnKeyAutomatically = true
-        $0.addTarget(self, action: #selector(tfDidChange), for: .editingChanged)
-    }
-
-    let pwBorderLine = UIView().then{
-        $0.backgroundColor = .todoaryGrey
-    }
-    
-    let pwNoticeLb = UILabel().then{
-        $0.isHidden = true
-        $0.text = "*영문, 숫자 포함 8자리 이상 입력해주세요 "
-        $0.labelTypeSetting(type: .sub1)
-        $0.textColor = .noticeRed
-        $0.font = UIFont.nbFont(type: .sub1)
-    }
-
-    let pwCertificationTf = UITextField().then{
-        $0.placeholder = "비밀번호 재입력"
-        $0.setPlaceholderColor()
-        $0.isSecureTextEntry = true
-        $0.font = UIFont.nbFont(type: .body2)
-        $0.textFieldTypeSetting(type: .body1)
-        $0.returnKeyType = .done
-        $0.enablesReturnKeyAutomatically = true
-        $0.addTarget(self, action: #selector(tfDidChange), for: .editingChanged)
-    }
-
-    let pwCertificationBorderLine = UIView().then{
-        $0.backgroundColor = .todoaryGrey
-    }
-    
-    let pwCertificationNoticeLb = UILabel().then{
-        $0.isHidden = true
-        $0.text = "*비밀번호가 일치하지 않습니다 "
-        $0.labelTypeSetting(type: .sub1)
-        $0.textColor = .noticeRed
-        $0.font = UIFont.nbFont(type: .sub1)
-    }
-
-
-    let confirmBtn = UIButton().then{
-        $0.isEnabled = false
-        $0.setTitle("확인", for: .normal)
-        $0.backgroundColor = .buttonColor
-        $0.setTitleColor(.white, for: .normal)
-        $0.titleLabel?.font = UIFont.nbFont(type: .button1)
-        $0.layer.cornerRadius = 52/2
-        $0.addTarget(self, action: #selector(passWordChangeBtnDidTap), for: .touchUpInside)
-    }
+    let mainView = PwFindView()
     
     
     //MARK: - Lifecycles
     
     override func viewDidLoad() {
-        super.viewDidLoad()
         
-        navigationTitle.text = "비밀번호 재설정"
+        super.viewDidLoad()
         self.view.backgroundColor = .white
         
-        self.idTf.delegate = self
-        self.certificationTf.delegate = self
-        self.pwTf.delegate = self
-        self.pwCertificationTf.delegate = self
-
-
-        setUpView()
-        setUpConstraint()
     }
+    
+    override func style(){
+        super.style()
+        navigationTitle.text = "비밀번호 재설정"
+    }
+    
+    override func layout(){
+        
+        super.layout()
+        
+        self.view.addSubview(mainView)
+        
+        mainView.snp.makeConstraints{
+            $0.top.equalToSuperview().offset(Const.Offset.top)
+            $0.leading.trailing.bottom.equalToSuperview()
+        }
+    }
+    
+    override func initialize() {
+        mainView.idTf.delegate = self
+        mainView.certificationTf.delegate = self
+        mainView.pwTf.delegate = self
+        mainView.pwCertificationTf.delegate = self
+        
+        mainView.idCertificationBtn.addTarget(self, action: #selector(idCertificationBtnDidTap), for: .touchUpInside)
+        mainView.certificationOkBtn.addTarget(self, action: #selector(certificationOkBtnDidTap), for: .touchUpInside)
+        mainView.confirmBtn.addTarget(self, action: #selector(passWordChangeBtnDidTap), for: .touchUpInside)
+        
+        mainView.pwTf.addTarget(self, action: #selector(tfDidChange), for: .editingChanged)
+        mainView.pwCertificationTf.addTarget(self, action: #selector(tfDidChange), for: .editingChanged)
+        
+    }
+    
+    //MARK: - Actions
     
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         self.view.endEditing(true)
@@ -198,8 +91,6 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
             self.view.window?.frame.origin.y = 0
         }
     }
-    
-    //MARK: - Actions
     
     @objc func backBtnDidTab() {
         self.navigationController?.popViewController(animated: true)
@@ -209,28 +100,28 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
         let text = sender.text ?? ""
         
         switch sender{
-        case idTf:
+        case mainView.idTf:
             self.email = text
-        case certificationTf:
+        case mainView.certificationTf:
             self.certification = text
-        case pwTf:
-            pwNoticeLb.isHidden = false
+        case mainView.pwTf:
+            mainView.pwNoticeLb.isHidden = false
             self.isValidPw = text.isValidPw()
             self.password = text
             if isValidPw{
-                pwNoticeLb.isHidden = true
+                mainView.pwNoticeLb.isHidden = true
                 }
             else {
-                pwNoticeLb.isHidden = false
+                mainView.pwNoticeLb.isHidden = false
             }
-        case pwCertificationTf:
-            pwCertificationNoticeLb.isHidden = false
+        case mainView.pwCertificationTf:
+            mainView.pwCertificationNoticeLb.isHidden = false
             if text == self.password {
                 isValidPwCheck = true
-                pwCertificationNoticeLb.isHidden = true
+                mainView.pwCertificationNoticeLb.isHidden = true
             }else{
                 isValidPwCheck = false
-                pwCertificationNoticeLb.isHidden = false
+                mainView.pwCertificationNoticeLb.isHidden = false
             }
         default:
             fatalError("Missing TextField...")
@@ -239,7 +130,7 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
     
 
     @objc func idCertificationBtnDidTap() {
-        self.email = idTf.text!
+        self.email = mainView.idTf.text!
         EmailCheckDataManager().emailCheckDataManager(self, email: self.email)
         print(self.email)
         print("버튼")
@@ -247,14 +138,14 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
     
     @objc func passWordChangeBtnDidTap(){
         
-        let pwInput = PwFindInput(email:idTf.text!, newPassword: self.password)
+        let pwInput = PwFindInput(email:mainView.idTf.text!, newPassword: self.password)
         
         PwFindDataManager().pwFindDataManager(self, pwInput)
 
     }
     
     @objc func certificationOkBtnDidTap(){
-        self.certification = certificationTf.text!
+        self.certification = mainView.certificationTf.text!
         if self.certification == UserDefaults.standard.string(forKey: "key"){
             isCertification = true
         }else{
@@ -278,14 +169,14 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
     //MARK: - Helpers
     
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        if textField == idTf {
-            certificationTf.becomeFirstResponder()
-            } else if textField == certificationTf {
-                pwTf.becomeFirstResponder()
-            }else if textField == pwTf {
-                pwCertificationTf.becomeFirstResponder()
-            }else if textField == pwCertificationTf {
-                pwCertificationTf.resignFirstResponder()
+        if textField == mainView.idTf {
+            mainView.certificationTf.becomeFirstResponder()
+        } else if textField == mainView.certificationTf {
+                mainView.pwTf.becomeFirstResponder()
+        }else if textField == mainView.pwTf {
+                mainView.pwCertificationTf.becomeFirstResponder()
+            }else if textField == mainView.pwCertificationTf {
+                mainView.pwCertificationTf.resignFirstResponder()
             }
             return true
         }
@@ -295,9 +186,9 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
             && isValidPwCheck
             && emailCheck
             && isCertification{
-            confirmBtn.isEnabled = true
+            mainView.confirmBtn.isEnabled = true
         }else{
-            confirmBtn.isEnabled = false
+            mainView.confirmBtn.isEnabled = false
         }
     }
     
@@ -308,14 +199,14 @@ class PwFindViewController: BaseViewController, UITextFieldDelegate {
             
             emailCheck = true
             
-            idNoticeLb.text = "*유효한 이메일입니다."
-            idNoticeLb.textColor = .todoaryGrey
+            mainView.idNoticeLb.text = "*유효한 이메일입니다."
+            mainView.idNoticeLb.textColor = .todoaryGrey
         
             MailSender.shared.sendEmail(email: self.email, viewController: self)
             
         }else if(code == 2017){
-            idNoticeLb.text = "*유효하지 않은 이메일입니다. 다시 입력해 주세요."
-            idNoticeLb.textColor = .noticeRed
+            mainView.idNoticeLb.text = "*유효하지 않은 이메일입니다. 다시 입력해 주세요."
+            mainView.idNoticeLb.textColor = .noticeRed
             emailCheck = false
         }else{
             emailCheck = false
